Put all classes into duke package
