+++ src/main/java/duke/MarkCommand.java
@@ -1,15 +1,17 @@
-public class UnmarkCommand extends Command {
+package duke;
+
+public class MarkCommand extends Command {
     public String input;
 
-    public UnmarkCommand(String input) {
+    public MarkCommand(String input) {
         this.input = input;
     }
 
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
         try {
             int index = Integer.parseInt(input.trim()) - 1;
-            taskList.get(index).markAsUndone();
-            ui.showUnmarkedTask(taskList.get(index));
+            taskList.get(index).markAsDone();
+            ui.showMarkedTask(taskList.get(index));
         } catch (NumberFormatException e) {
             throw new DukeException("Input a valid number!");
         }
