+++ src/main/java/duke/UnmarkCommand.java
@@ -1,15 +1,17 @@
-public class MarkCommand extends Command {
+package duke;
+
+public class UnmarkCommand extends Command {
     public String input;
 
-    public MarkCommand(String input) {
+    public UnmarkCommand(String input) {
         this.input = input;
     }
 
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
         try {
             int index = Integer.parseInt(input.trim()) - 1;
-            taskList.get(index).markAsDone();
-            ui.showMarkedTask(taskList.get(index));
+            taskList.get(index).markAsUndone();
+            ui.showUnmarkedTask(taskList.get(index));
         } catch (NumberFormatException e) {
             throw new DukeException("Input a valid number!");
         }
